+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Features:
 * Support Python 3.11.
 * Support `typing_extensions.overload` and `typing_extensions.NamedTuple`. (The latter
   is not yet released, but will be in the next version of `typing_extensions`.)
+* Extend Y001 check to cover type aliases as well as `TypeVar`s, `ParamSpec`s and
+  `TypeVarTuple`s.
 
 ## 22.5.1